requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
tabula-py==2.9.0
python-docx==1.0.1
numpy>=1.19.0
pdf2docx==0.5.6 